refactor: 각 Page 의 Item 컴포넌트 depth 변경
