--- src/pages/CartPage/CartProductItem/style.js
@@ -1,35 +1,0 @@
-import styled from 'styled-components';
-import { PATTERN_FILE_NAME, COLOR } from '../../../constants';
-
-export const Container = styled.li`
-  padding: 1.5rem 0.25rem;
-  box-sizing: border-box;
-  display: flex;
-`;
-
-export const Image = styled.img.attrs((props) => ({
-  src: props.src,
-  alt: props.src.match(PATTERN_FILE_NAME)[0],
-}))`
-  width: 9rem;
-`;
-
-export const Name = styled.span`
-  flex-grow: 1;
-  padding-left: 1.25rem;
-  font-size: 1.25rem;
-  color: ${COLOR.HEX.GRAY_800};
-`;
-
-export const Controller = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  align-items: flex-end;
-`;
-
-export const Price = styled.span`
-  font-size: 1rem;
-  color: ${COLOR.HEX.GRAY_800};
-  text-align: right;
-`;
