+++ src/pages/CheckoutPage/style.js
@@ -1,6 +1,7 @@
 import styled from 'styled-components';
+
 import { CheckoutBox, List, Template } from '../../components';
-import { COLOR } from '../../constants';
+import { COLOR, PATTERN_FILE_NAME } from '../../constants';
 
 export const Page = styled(Template)`
   background-color: ${COLOR.HEX.WHITE};
@@ -49,4 +50,36 @@ export const StickyCheckoutBox = styled(CheckoutBox)`
   top: 15%;
   margin-top: 4.875rem;
   width: 100%;
+`;
+
+/* Item */
+
+export const Container = styled.li`
+  display: flex;
+  padding: 1.5rem 1.5rem;
+  box-sizing: border-box;
+`;
+
+export const Content = styled.div`
+  display: flex;
+  flex-direction: column;
+  padding-left: 1.25rem;
+`;
+
+export const Name = styled.span`
+  font-size: 1.25rem;
+  color: ${COLOR.HEX.GRAY_800};
+`;
+
+export const Quantity = styled.span`
+  font-size: 1rem;
+  padding-top: 1.5rem;
+  color: ${COLOR.HEX.GRAY_800};
+`;
+
+export const Image = styled.img.attrs((props) => ({
+  src: props.src,
+  alt: props.src.match(PATTERN_FILE_NAME)[0],
+}))`
+  width: 7.5rem;
 `;
