+++ src/pages/index.js
@@ -1,4 +1,5 @@
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
+
 import { ProductListPage } from './ProductListPage';
 import { CartPage } from './CartPage';
 import { CheckoutPage } from './CheckoutPage';
